requirements.txt
numpy>=1.24
pandas>=2.0
Pillow>=10.0
reportlab>=4.0
mpi4py>=3.1

# Optional for real microscopy and sparse scientific workflows:
# scipy
# scikit-image
# cellpose
